addBook finished, needs error handling for closing scanner

+++ projectLibrary/src/models/Book.java
@@ -5,12 +5,12 @@ import java.util.Objects;
 public class Book {
     private int id;
     private String title;
-    private float price;
+    private double price;
 
     public Book() {
     }
 
-    public Book(int id, String title, float price) {
+    public Book(int id, String title, double price) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -33,11 +33,11 @@ public class Book {
         this.title = title;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
@@ -47,7 +47,7 @@ public class Book {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Book book = (Book) o;
-        return id == book.id && Float.compare(book.price, price) == 0 && Objects.equals(title, book.title);
+        return id == book.id && Double.compare(book.price, price) == 0 && Objects.equals(title, book.title);
     }
 
     @Override
